Fix `LoadButton` stub jumping
