+++ src/_locales/ru/messages.json
@@ -6,19 +6,19 @@
     "message": "Для скачивания видео перейдите на страницу сериала или фильма на сайте"
   },
   "popup_stub_trailer": {
-    "message": "Похоже что контент в ожидании выхода.\nВернитесь сюда позже!"
+    "message": "Похоже, что контент в ожидании выхода.\nВернитесь сюда позже!"
   },
   "popup_stub_locationFilm": {
-    "message": "Увы, но фильм не доступен с вашего местоположения :("
+    "message": "К сожалению, этот фильм не доступен в вашем регионе :("
   },
   "popup_stub_locationSerial": {
-    "message": "Увы, но сериал не доступен с вашего местоположения :("
+    "message": "К сожалению, этот сериал не доступен в вашем регионе :("
   },
   "popup_stub_vpn": {
-    "message": "Попробуйте воспользоваться VPN."
+    "message": "Попробуйте воспользоваться VPN"
   },
   "popup_stub_unavailable": {
-    "message": "Контент недоступен для загрузки — отсутствуют данные для обработки."
+    "message": "Контент недоступен для загрузки — отсутствуют данные для обработки"
   },
   "popup_stub_error": {
     "message": "Упс! Кажется что-то пошло не так..."
