+++ src/_locales/uk/messages.json
@@ -9,16 +9,16 @@
     "message": "Схоже, що контент в очікуванні виходу.\nПоверніться сюди пізніше!"
   },
   "popup_stub_locationFilm": {
-    "message": "На жаль, але фільм недоступний з вашого місцезнаходження :("
+    "message": "На жаль, але фільм недоступний у вашому регіоні :("
   },
   "popup_stub_locationSerial": {
-    "message": "На жаль, але серіал недоступний з вашого місцезнаходження :("
+    "message": "На жаль, але серіал недоступний у вашому регіоні :("
   },
   "popup_stub_vpn": {
-    "message": "Спробуйте скористатися VPN."
+    "message": "Спробуйте скористатися VPN"
   },
   "popup_stub_unavailable": {
-    "message": "Контент недоступний для завантаження — відсутні дані для обробки."
+    "message": "Контент недоступний для завантаження — відсутні дані для обробки"
   },
   "popup_stub_error": {
     "message": "Упс! Здається, щось пішло не так..."
